Add function for transforming Frame
This allows us to transform group of paths and/or texts without specifying transform for every one of them.
Local transforms that are specified in paths/texts get premultiplied by this transform.
This is nice when you have function that draws one component/widget composed of multiple paths and you want to
treat that function as one draw call and transform it accordingly, without the calee of the function worrying about anything.