Remove default features and bump to 1.0.0
Crate is ready for publish.
@KevinKelley please publish with `cargo publish --no-verify`
for nanovg and nanovg-sys.
I manually built the crates with all features (debug + release),
everything built and ran fine on my machine.

Regarding 1.0.0:
Our crate seems to be mature enough (features and stability)
to release a first full-semver version. This doesn't necessarily mean
that the crate is "finished" or 100% correct and bug free, but it
signals to users that they can use it without fear; it also makes
it a bit easier to work with, due to the proper semver versioning.

Regarding removal of the default features:
This forces the user to specify which OpenGL version to use and thus
removes our assumption on the GL capabilities of their machine.
While it's true that default features make it easier for beginners,
I think it would be way worse if a beginner tried to build the crate,
with the default feature (gl3) enabled, on a machine that only supports
gl2, as an example. Debugging that problem would be way harder than
understanding our "Have you enabled a feature?" message in build.rs.